backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
charset-normalizer==3.4.4
click==8.1.8
exceptiongroup==1.3.1
fastapi==0.128.8
filelock==3.19.1
fsspec==2025.10.0
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
packaging==26.0
pillow==11.3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
python-multipart==0.0.20
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
safetensors==0.7.0
sse-starlette==3.3.0
starlette==0.49.3
sympy==1.14.0
tokenizers==0.22.2
torch==2.8.0
tqdm==4.67.3
transformers==4.57.6
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.39.0
